:recycle: move model path classes to serving_models
